Move extensions into Domain
