+++ src/Employer/Employer.Web/Mappings/VacancyDisplayViewModelMapper.cs
@@ -4,6 +4,7 @@ using Esfa.Recruit.Employer.Web.ViewModels;
 using Esfa.Recruit.Shared.Web.Extensions;
 using Esfa.Recruit.Shared.Web.Services;
 using Esfa.Recruit.Vacancies.Client.Domain.Entities;
+using Esfa.Recruit.Vacancies.Client.Domain.Extensions;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.Client;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.Services.Wages;
 using Humanizer;
@@ -49,7 +50,7 @@ namespace Esfa.Recruit.Employer.Web.Mappings
             vm.ContactName = vacancy.EmployerContactName;
             vm.ContactEmail = vacancy.EmployerContactEmail;
             vm.ContactTelephone = vacancy.EmployerContactPhone;
-            vm.ClosingDate = vacancy.ClosingDate?.AsDisplayDate();
+            vm.ClosingDate = vacancy.ClosingDate?.AsGdsDate();
             vm.EmployerDescription = vacancy.EmployerDescription;
             vm.EmployerName = vacancy.EmployerName;
             vm.EmployerWebsiteUrl = vacancy.EmployerWebsiteUrl;
@@ -60,7 +61,7 @@ namespace Esfa.Recruit.Employer.Web.Mappings
                 ? $"{"position".ToQuantity(vacancy.NumberOfPositions.Value)} available"
                 : null;
             vm.OutcomeDescription = vacancy.OutcomeDescription;
-            vm.PossibleStartDate = vacancy.StartDate?.AsDisplayDate();
+            vm.PossibleStartDate = vacancy.StartDate?.AsGdsDate();
             vm.ProviderName = vacancy.TrainingProvider?.Name;
             vm.Qualifications = vacancy.Qualifications.SortQualifications(_qualifications.Value).AsText();
             vm.ShortDescription = vacancy.ShortDescription;
