+++ src/Employer/Employer.Web/Views/_ViewImports.cshtml
@@ -5,6 +5,7 @@
 @using Esfa.Recruit.Employer.Web.Views
 @using Esfa.Recruit.Employer.Web.Extensions
 @using Esfa.Recruit.Shared.Web.Extensions
+@using Esfa.Recruit.Vacancies.Client.Domain.Extensions
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Recruit.Shared.Web
 @addTagHelper *, Esfa.Recruit.Employer.Web
