--- src/Shared/Recruit.Shared.Web/Extensions/DateTimeExtensions.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace Esfa.Recruit.Shared.Web.Extensions
-{
-    public static class DateTimeExtensions
-    {
-        private const string DisplayDateFormat = "dd MMM yyyy";
-        
-        public static string AsDisplayDate(this DateTime date)
-        {
-            return date.ToString(DisplayDateFormat);
-        }
-    }
-}
