+++ src/Shared/Recruit.Vacancies.Client/Domain/Extensions/DecimalExtensions.cs
@@ -1,4 +1,4 @@
-namespace Esfa.Recruit.Shared.Web.Extensions
+namespace Esfa.Recruit.Vacancies.Client.Domain.Extensions
 {
     public static class DecimalExtensions
     {
